requirements.txt
streamlit
python-binance
pandas
ta
requests
plotly
streamlit_js_eval